WIP for merge into gstreamer

+++ neuralet-distancing.py
@@ -1,62 +1,87 @@
 #!/usr/bin/python3
 import argparse
-from multiprocessing import Process
-import threading
-from libs.config_engine import ConfigEngine
 import logging
+import sys
+import threading
+
+from multiprocessing import Process
+from libs.config_engine import ConfigEngine
+
 
 logger = logging.getLogger(__name__)
 
 
-def start_engine(config, video_path):
-    import pdb, traceback, sys
+def start_cv_engine(config, video_path):
     try:
         if video_path:
             from libs.core import Distancing as CvEngine
             engine = CvEngine(config)
             engine.process_video(video_path)
         else:
-            logger.info('Skipping CVEngine as video_path is not set')
-    except:
-        extype, value, tb = sys.exc_info()
-        traceback.print_exc()
-        pdb.post_mortem(tb)
+            logger.error('"VideoPath" not set in .ini [App] section')
+    except Exception:
+        # this runs sys.excinfo() and logs the result
+        logger.error("CvEngine failed.", exc_info=True)
 
 def start_web_gui(config):
     from ui.web_gui import WebGUI
     ui = WebGUI(config)
     ui.start()
 
-
-def main(config):
-    logging.basicConfig(level=logging.INFO)
+def main(config, verbose=False):
+    logging.basicConfig(level=logging.DEBUG if verbose else logging.INFO)
     if isinstance(config, str):
         config = ConfigEngine(config)
-
     video_path = config.get_section_dict("App").get("VideoPath", None)
-    process_engine = Process(target=start_engine, args=(config, video_path,))
+
+    # create our inference process
+    try:
+        # try to launch deepstream but skip if any import errors
+        # (eg. pyds not found, gi not found)
+        from libs.detectors.deepstream import DsEngine, DsConfig
+        process_engine = DsEngine(DsConfig(config))
+    except ImportError:
+        # DeepStream is not available. Let's try CvEngine
+        process_engine = Process(target=start_cv_engine, args=(config, video_path,))
+
+    # create our ui process
     process_api = Process(target=start_web_gui, args=(config,))
 
+    # start both processes
     process_api.start()
     process_engine.start()
     logger.info("Services Started.")
 
+    # wait forever
     forever = threading.Event()
     try:
         forever.wait()
     except KeyboardInterrupt:
         logger.info("Received interrupt. Terminating...")
 
-    process_engine.terminate()
-    process_engine.join()
-    logger.info("CV Engine terminated.")
+    if hasattr(process_engine, 'stop'):
+        # DsEngine shuts down by asking
+        # GLib.MainLoop to quit. SIGTERM does this too,
+        # but it wouldn't call some extra debug stuff
+        # that's in GstEngine's quit()
+        # (debug .dot file, .pdf if graphviz is available)
+        # .stop() will call .terminate() if it times out.
+        process_engine.stop()
+        process_engine.join()
+    else:
+        process_engine.terminate()
+        process_engine.join()
+
+    logger.info("Inference Engine terminated.")
     process_api.terminate()
     process_api.join()
     logger.info("Web GUI terminated.")
+    return 0
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('--config', required=True)
+    parser.add_argument('--verbose', action='store_true')
     args = parser.parse_args()
-    main(args.config)
+    sys.exit(main(args.config, args.verbose))
